fix(cli): tsconfig with files and include entries fix #860
